cosmetics.
svn:sync-xref-src-repo-rev: r28036

+++ src/VBox/Devices/PC/DevPcBios.cpp
@@ -1,4 +1,4 @@
-/* $Id: DevPcBios.cpp 6876 2008-02-10 19:44:59Z noreply@oracle.com $ */
+/* $Id: DevPcBios.cpp 6890 2008-02-11 10:47:39Z knut.osmundsen@oracle.com $ */
 /** @file
  * PC BIOS Device.
  */
@@ -843,7 +843,7 @@ static int pcbiosPlantDMITable(PPDMDEVINS pDevIns, uint8_t *pTable, unsigned cbM
 #define STRCPY(p, s) \
     do { \
         size_t _len = strlen(s) + 1; \
-        size_t _max = (unsigned)(pszStr + _len - (char *)pTable) + 1; \
+        size_t _max = (size_t)(pszStr + _len - (char *)pTable) + 1; /* +1 for strtab terminator */ \
         if (_max > cbMax) \
             return PDMDevHlpVMSetError(pDevIns, VERR_TOO_MUCH_DATA, RT_SRC_POS, \
                     N_("One of the DMI strings is too long. Check all bios/Dmi* configuration entries. At least %zu bytes are needed but there is no space for more than %d bytes"), _max, cbMax); \
